# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "MYPLY SERVER",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "minkj1992@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/memos/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "내 메모 리스트 조회",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memos"
                ],
                "summary": "Get user's Memo list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ListMemoResponse"
                        }
                    },
                    "401": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "메모 생성",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memos"
                ],
                "summary": "Add Memo",
                "parameters": [
                    {
                        "description": "memo request body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.AddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.MemoResponse"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/memos/{memoID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "메모 조회",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memos"
                ],
                "summary": "Retrieve Memo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "memoID to retrieve",
                        "name": "memoID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.MemoResponse"
                        }
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "메모 수정",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memos"
                ],
                "summary": "Update Memo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "memoID to retrieve",
                        "name": "memoID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "memo request body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.PatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.MemoResponse"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/memos/{memoIdOrYoutubeID}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "1. (Device Token, memo ID) 조합으로 메모 삭제\n2. (Device Token, YoutubeVideoID) 조합으로 메모 삭제",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memos"
                ],
                "summary": "Delete Memo by a unique id (memo ID or youtube video ID)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "memoIdOrYoutubeID to get a memo",
                        "name": "memoIdOrYoutubeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.AddRequest": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "youtubeVideoID": {
                    "type": "string"
                }
            }
        },
...
        "controller.ListMemoData": {
            "type": "object",
            "properties": {
                "memos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.MemoResponse"
                    }
                }
            }
        },
        "controller.ListMemoResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/controller.ListMemoData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "controller.MemoResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "memoID": {
                    "type": "string"
                },
                "thumbnailURL": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "youtubeVideoID": {
                    "type": "string"
                }
            }
        },
...
        "controller.PatchRequest": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Device-Token",
            "in": "header"
        }
    }
}